feat: add xml importer v2
